fix: movido servicos sem vinculo com endpoints para a classe sua respectiva classe util
